Update `program.py` to workaround an MSPFlasher issue
MSPFlasher seems to silently ignore the second read command in a sequence, so we instead now do one big read command that dumps the full range that includes all the bytes we need rather than dumping the specific smaller ranges. Then we parse out the bytes we need from the resulting dump. More complicated and dumps more bytes than we need, but works.