Disable `RESET` pin between programming
Some units programmed at CW&T seemed to have bad chips, so maybe this reset pin is messing up the chip when it contacts the `Vcc` pin as the PCB is slid into the programming jig? Disabling the `RESET` does not seem to cause any problems with the chip booting up after programming, so mind as well try this and see if it has an effect.