Disable backup switchover in the RTC config
Count power usage:

Prev firmware with backup enabled and cap on PCB | 2.0uA
New firmware with cap on PCB (should not happen) | 2.0uA
New firmware with 10K resistor in place C2 | 2.0uA

...and the failure case where old firmware gets on new PCB | 120uA

So we need to make sure that old firmware does not get on new PCBs. It should get caught during commissioning because the power will not be able to ride the time between programming and battery insertion.